# kogi-infra/openapi/kogi-server.yaml
openapi: 3.0.3
info:
  title: Kogi MVP Server API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/database/query:
    post:
      summary: Execute a database query via service facade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sql:
                  type: string
      responses:
        "200":
          description: database query result
...
  /api/v1/auth/provision:
    post:
      summary: Trigger provisioner flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                plan:
                  type: string
      responses:
        "202":
          description: provisioning queued
...